Fix Visual Studio build (/wX)
warning C4800: 'uint32_t': forcing value to bool 'true' or 'false' (performance warning)
warning C4091: 'typedef ': ignored on left of '' when no variable is declared
warning C4996: 'chdir': The POSIX name for this item is deprecated.
etc

+++ dep/libmpq/libmpq/mpq_libmpq04.h
@@ -57,7 +57,7 @@ public:
             token[strlen(token) - 1] = 0;
             std::string s = token;
             filelist.push_back(s);
-            counter += strlen(token) + 2;
+            counter += static_cast<uint32_t>(strlen(token) + 2);
             token = strtok(NULL, seps);
         }
